+++ failfast/src/main/kotlin/failfast/FailFastDSL.kt
@@ -1,0 +1,4 @@
+package failfast
+
+@DslMarker
+annotation class FailFastDSL