+++ failfast/src/main/kotlin/failfast/TestDSL.kt
@@ -1,3 +1,4 @@
 package failfast
 
+@FailFastDSL
 class TestDSL